dev-requirements.txt
mock
nose
nose-cov
testtools
testfixtures
coverage==3.7.1
pyldap